+++ lib/interfaces/IAssertion.ts
@@ -18,6 +18,7 @@ interface IEqualityAssertion {
   toStrictNotEqual(expected: any): IAssertion;
   toBe(expected: any): IAssertion;
   toNotBe(expected: any): IAssertion;
+  toObjectEqual(expected: any): IAssertion;
 }
 
 interface ITruthinessAssertion {