+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## Added:
+* New assertions. See [assertions API](README.md#assertions-api) 
+
 ## Fix:
 * Fixed error on wrong keybinding pressed
 * Updated module cache deleting. If function were changed, watch mode reruns tests for new implementation